requirements.txt
dj-database-url==0.5.0
Django==2.2.6
djangorestframework==3.10.3
gunicorn==19.9.0
psycopg2==2.8.3
pytz==2019.3
sqlparse==0.3.0
